Added GNU GPL
s.t. anybody can add and modify these works, and if they've modified considerably they should be able to sell, or if they've distributed and as such they charge for it. Couldve gone with apache, or maybe one that doesnt allow any selling but in any case the idea is to allow people to contribute to a project and then those contributions always remain free to use for anyone. If someone makes changes and makes it better, they shouldnt be able to charge for everything else that was done by everyone else. Like this is too complicated - so lets stick to GPL and help everyone plan their todos better